Inputs should be floats, not ints

+++ row.py
@@ -1,4 +1,4 @@
 class row:
-    def __init__(self, selectionClass: int, selectionData: list[int]) -> None:
+    def __init__(self, selectionClass: int, selectionData: list[float]) -> None:
         self.selectionClass = selectionClass
         self.selectionData = selectionData
